move sql readers to sql.reader, add hook in burdock.sql. Make API across reader classes more uniform
