--- sdk/burdock/reader/sql/__init__.py
@@ -1,12 +1,0 @@
-
-from burdock.reader.sql.pandas import PandasReader
-from burdock.reader.sql.presto import PrestoReader
-from burdock.reader.sql.postgres import PostgresReader
-from burdock.reader.sql.sql_server import SqlServerReader
-from burdock.reader.sql.spark import SparkReader
-
-__all__ = ["PandasReader",
-           "PostgresReader",
-           "PrestoReader",
-           "SqlServerReader",
-           "SparkReader"]
