+++ tests/query/test_interval_expr.py
@@ -5,8 +5,7 @@ import pandas as pd
 from pandasql import sqldf
 
 from burdock.metadata import CollectionMetadata
-from burdock.sql import PrivateReader
-from burdock.reader.sql import PandasReader
+from burdock.sql import PrivateReader, PandasReader
 from burdock.sql.parse import QueryParser
 from burdock.reader.rowset import TypedRowset
 
@@ -23,7 +22,7 @@ df = pd.read_csv(csv_path)
 class TestQuery:
     def test_group_by_noisy_typed_order_inter(self):
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 1.0)
+        private_reader = PrivateReader(schema, reader, 1.0)
         rs = private_reader.execute_typed("SELECT COUNT(*) AS c, married AS m FROM PUMS.PUMS GROUP BY married ORDER BY c")
         assert(rs['c'][0] < rs['c'][1])
         assert(len(rs.report['c'].intervals[0.95]) == 2)
@@ -34,7 +33,7 @@ class TestQuery:
         assert(all(outer.high > inner.high for inner, outer in zip(rs.report['c'].intervals[0.95], rs.report['c'].intervals[0.985])))
     def test_group_by_noisy_typed_order_inter_constant(self):
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 1.0)
+        private_reader = PrivateReader(schema, reader, 1.0)
         rs = private_reader.execute_typed("SELECT COUNT(*) AS c, married AS m FROM PUMS.PUMS GROUP BY married ORDER BY c")
         rs2 = private_reader.execute_typed("SELECT COUNT(*) * 2 AS c, married AS m FROM PUMS.PUMS GROUP BY married ORDER BY c")
         assert(len(rs.report['c'].intervals[0.95]) == 2)
