+++ tests/query/test_query.py
@@ -5,8 +5,7 @@ import pandas as pd
 from pandasql import sqldf
 
 from burdock.metadata import CollectionMetadata
-from burdock.sql import PrivateReader
-from burdock.reader.sql import PandasReader
+from burdock.sql import PrivateReader, PandasReader
 from burdock.sql.parse import QueryParser
 from burdock.reader.rowset import TypedRowset
 
@@ -51,42 +50,42 @@ class TestQuery:
         assert(rs[2][0] == 451 * 5)
     def test_group_by_noisy_order(self):
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 1.0)
+        private_reader = PrivateReader(schema, reader, 1.0)
         rs = private_reader.execute("SELECT COUNT(*) AS c, married AS m FROM PUMS.PUMS GROUP BY married ORDER BY c")
         assert(rs[1][0] < rs[2][0])
     def test_group_by_noisy_order_desc(self):
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 1.0)
+        private_reader = PrivateReader(schema, reader, 1.0)
         rs = private_reader.execute("SELECT COUNT(*) AS c, married AS m FROM PUMS.PUMS GROUP BY married ORDER BY c DESC")
         assert(rs[1][0] > rs[2][0])
     def test_group_by_noisy_typed_order(self):
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 1.0)
+        private_reader = PrivateReader(schema, reader, 1.0)
         rs = private_reader.execute_typed("SELECT COUNT(*) AS c, married AS m FROM PUMS.PUMS GROUP BY married ORDER BY c")
         assert(rs['c'][0] < rs['c'][1])
     def test_group_by_noisy_typed_order_desc(self):
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 1.0)
+        private_reader = PrivateReader(schema, reader, 1.0)
         rs = private_reader.execute_typed("SELECT COUNT(*) AS c, married AS m FROM PUMS.PUMS GROUP BY married ORDER BY c DESC")
         assert(rs['c'][0] > rs['c'][1])
     def test_no_tau(self):
         # should never drop rows
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 4.0)
+        private_reader = PrivateReader(schema, reader, 4.0)
         for i in range(10):
             rs = private_reader.execute_typed("SELECT COUNT(*) AS c FROM PUMS.PUMS WHERE age > 90 AND educ = '8'")
             assert(len(rs['c']) == 1)
     def test_no_tau_noisy(self):
         # should never drop rows
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 0.01)
+        private_reader = PrivateReader(schema, reader, 0.01)
         for i in range(10):
             rs = private_reader.execute_typed("SELECT COUNT(*) AS c FROM PUMS.PUMS WHERE age > 90 AND educ = '8'")
             assert(len(rs['c']) == 1)
     def test_yes_tau(self):
         # should usually drop some rows
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 1.0, 1/10000)
+        private_reader = PrivateReader(schema, reader, 1.0, 1/10000)
         lengths = []
         for i in range(10):
             rs = private_reader.execute_typed("SELECT COUNT(*) AS c FROM PUMS.PUMS WHERE age > 80 GROUP BY educ")
@@ -106,7 +105,7 @@ class TestQuery:
     def test_empty_result_count_typed_notau_prepost(self):
         reader = PandasReader(schema, df)
         query = QueryParser(schema).queries("SELECT COUNT(*) as c FROM PUMS.PUMS WHERE age > 100")[0]
-        private_reader = PrivateReader(reader, schema, 1.0)
+        private_reader = PrivateReader(schema, reader, 1.0)
         private_reader._execute_ast(query, True)
         for i in range(3):
             trs = private_reader._execute_ast(query, True)
@@ -118,6 +117,6 @@ class TestQuery:
         assert(trs['age_total'][0] > 1000)
     def test_sum_noisy_postprocess(self):
         reader = PandasReader(schema, df)
-        private_reader = PrivateReader(reader, schema, 1.0)
+        private_reader = PrivateReader(schema, reader, 1.0)
         trs = private_reader.execute_typed("SELECT POWER(SUM(age), 2) as age_total FROM PUMS.PUMS")
         assert(trs['age_total'][0] > 1000 ** 2)
