+++ tests/reader/test_reader_base.py
@@ -1,8 +1,8 @@
 import pytest
 
 from burdock.reader.base import Reader
-from burdock.reader.sql.engine import Engine
-from burdock.reader.sql import PandasReader, PrestoReader, PostgresReader, SqlServerReader, SparkReader
+from burdock.sql.reader.engine import Engine
+from burdock.sql.reader import PandasReader, PrestoReader, PostgresReader, SqlServerReader, SparkReader
 
 READERS = PandasReader, PrestoReader, PostgresReader, SqlServerReader, SparkReader
